add solid to the components shopping cart and cards
