+++ nextjs-dashboard/app/dashboard/menu/page.tsx
@@ -1,41 +1,11 @@
 'use client';
 
-import { useState } from 'react';
-import Card from '@/app/ui/menu-dishes/cards';
-import ShoppingCart from '@/app/ui/menu-shopping-card/shoppingCart';
-
-interface CartItem {
-  name: string;
-  quantity: number;
-}
+import Card from '@/app/ui/menu-dishes/components/cards';
+import ShoppingCart from '@/app/ui/menu-shopping-card/components/shoppingCart';
+import { useCart } from '@/app/ui/menu-dishes/hooks/useCart';
 
 export default function MenusPage() {
-  const [cartItems, setCartItems] = useState<CartItem[]>([]);
-  const [isCartVisible, setIsCartVisible] = useState(false); // Controlar la visibilidad del carrito
-
-  const addToCart = (name: string, quantity: number) => {
-    const existingItem = cartItems.find((item) => item.name === name);
-    if (existingItem) {
-      setCartItems(
-        cartItems.map((item) =>
-          item.name === name
-            ? { ...item, quantity: item.quantity + quantity }
-            : item
-        )
-      );
-    } else {
-      setCartItems([...cartItems, { name, quantity }]);
-    }
-    setIsCartVisible(true); // Mostrar el carrito cuando se añade un ítem
-  };
-
-  const removeFromCart = (name: string) => {
-    setCartItems(cartItems.filter((item) => item.name !== name));
-  };
-
-  const handleCloseCart = () => {
-    setIsCartVisible(false); // Cerrar el carrito
-  };
+  const { cartItems, addToCart, removeFromCart, isCartVisible, handleCloseCart } = useCart();
 
   return (
     <div className="p-8 grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-8">
@@ -58,12 +28,11 @@ export default function MenusPage() {
         addToCart={addToCart}
       />
 
-      {/* Menú de carrito lateral */}
       {isCartVisible && (
         <ShoppingCart
           cartItems={cartItems}
           removeFromCart={removeFromCart}
-          onClose={handleCloseCart} // Pasar la función para cerrar el carrito
+          onClose={handleCloseCart}
         />
       )}
     </div>
