+++ nextjs-dashboard/app/shared/interfaces/CartItem.ts
@@ -1,0 +1,4 @@
+export interface CartItem {
+    name: string;
+    quantity: number;
+}
